refactor: DnDCurrentType을 dnd.type으로 이동
