+++ frontend/techpick/src/types/dnd.type.ts
@@ -1,0 +1,10 @@
+import type { UniqueIdentifier } from '@dnd-kit/core';
+
+export type DnDCurrentType = {
+  id: UniqueIdentifier;
+  sortable: {
+    containerId: string | null;
+    items: UniqueIdentifier[];
+    index: number;
+  };
+};
