+++ frontend/techpick/src/types/index.ts
@@ -2,3 +2,4 @@ export type { NodeData, DirectoryNodeProps } from './NodeData';
 export type * from './tag.type';
 export type * from './folder.type';
 export type * from './pick.type';
+export type * from './dnd.type';
